Combined ChooseColor and ChooseLangScreens into ChooseColorLangScreen and and improve color option handling.
